+++ webapp-backend/src/main/java/org/oransc/ric/portal/dashboard/DashboardConstants.java
@@ -31,6 +31,8 @@ public abstract class DashboardConstants {
 
 	// Factor out method names used in multiple controllers
 	public static final String VERSION_METHOD = "version";
+	
+	public static final String APP_NAME_AC = "AC";
 
 	// The role names are defined by ONAP Portal.
 	// The prefix "ROLE_" is required by Spring.