+++ webapp-backend/src/test/java/org/oransc/ric/portal/dashboard/controller/AdminControllerTest.java
@@ -21,11 +21,16 @@ package org.oransc.ric.portal.dashboard.controller;
 
 import java.lang.invoke.MethodHandles;
 import java.net.URI;
+import java.util.HashMap;
 import java.util.List;
+import java.util.Map;
 
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
+import org.oransc.ric.portal.dashboard.DashboardConstants;
 import org.oransc.ric.portal.dashboard.model.DashboardUser;
+import org.oransc.ric.portal.dashboard.model.ErrorTransport;
+import org.oransc.ric.portal.dashboard.model.IDashboardResponse;
 import org.oransc.ric.portal.dashboard.model.SuccessTransport;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -36,6 +41,7 @@ import org.springframework.http.ResponseEntity;
 public class AdminControllerTest extends AbstractControllerTest {
 
 	private static final Logger logger = LoggerFactory.getLogger(MethodHandles.lookup().lookupClass());
+	//public static final String APP_NAME = DashboardConstants.APP_NAME;
 
 	@Test
 	public void versionTest() {
@@ -70,6 +76,31 @@ public class AdminControllerTest extends AbstractControllerTest {
 		ResponseEntity<String> response = testRestTemplateStandardRole().exchange(uri, HttpMethod.GET, null,
 				String.class);
 		Assertions.assertTrue(response.getStatusCode().is4xxClientError());
+	}
+
+	@Test
+	public void getxAppMetricsUrlTest() {
+		Map<String, String> metricsQueryParms = new HashMap<String, String>();
+		metricsQueryParms.put("app", DashboardConstants.APP_NAME_AC);
+		URI uri = buildUri(metricsQueryParms, AdminController.CONTROLLER_PATH, AdminController.XAPPMETRICS_METHOD);
+		logger.debug("Invoking {}", uri);
+		ResponseEntity<SuccessTransport> successResponse = testRestTemplateStandardRole().exchange(uri, HttpMethod.GET, null,
+				SuccessTransport.class);
+		Assertions.assertFalse(successResponse.getBody().getData().toString().isEmpty());
+		Assertions.assertTrue(successResponse.getStatusCode().is2xxSuccessful());
+	}
+
+	@Test
+	public void getxAppMetricsUrlTestFail() {
+		Map<String, String> metricsQueryParms = new HashMap<String, String>(); 
+		//Providing a bogus value for application name in query parameter to test failure 
+		metricsQueryParms.put("app", "ABCD");
+		URI uri = buildUri(metricsQueryParms, AdminController.CONTROLLER_PATH, AdminController.XAPPMETRICS_METHOD);
+		logger.debug("Invoking {}", uri);
+		ResponseEntity<ErrorTransport> errorResponse = testRestTemplateStandardRole().exchange(uri, HttpMethod.GET, null,
+				ErrorTransport.class);
+		logger.debug("{}", errorResponse.getBody().getError().toString());
+		Assertions.assertTrue(errorResponse.getStatusCode().is4xxClientError());
 	}
 
 }